WIP Enable users to derive extra params from params + site data
so that they can f.ex. display a list of tags with their counts.

Instead of adding `:extend-params-fn` to the overrides, we could pass it
as an extra param but it is complicated because we need to do it on 2
functions and would then perhaps need to support 3 arities. This seems
simpler.

* Rename `overrides` to something better (what???)
* Use Spec to document the arguments to the extend-params-fn?
* Document extend-params-fn in docstrings and perhaps the site docs?